pyproject.toml
[project]
name = "inframind-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "uuid6>=2025.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.10",
]

[tool.ruff]
exclude = [".venv","app/repository"]

[tool.ty.src]
exclude =[".venv"]
